refact: do file-structuring overall
